+++ src/test/java/com/epam/service/PersonServiceTest.java
@@ -1,63 +1,50 @@
 package com.epam.service;
 
+
 import com.epam.model.Person;
 import com.epam.model.PersonRoleEnum;
+import com.epam.repository.PersonDAO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.util.ArrayList;
+import java.util.List;
 
-import static org.mockito.Mockito.when;
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.when;
 
-
-class PersonDetailsServiceImplTest {
+class PersonServiceTest {
     @Mock
+    private PersonDAO personDAO;
+
+    private List<Person> personList;
     private PersonService personService;
-    @Mock
-    private Person person;
-
-    private Set<GrantedAuthority> roles;
-    private PersonDetailsServiceImpl personDetailsServiceImpl;
+    private Person expectedPerson;
 
     @BeforeEach
     void setUp() {
         MockitoAnnotations.initMocks(this);
-        personDetailsServiceImpl = new PersonDetailsServiceImpl(personService);
-        roles = new HashSet<>();
+        expectedPerson = new Person(1, "user", "1111", PersonRoleEnum.valueOf("ADMIN"));
+        personService = new PersonService(personDAO);
+        personList = new ArrayList<>();
+        personList.add(expectedPerson);
     }
 
     @Test
-    void loadUserPositiveResult() {
-        String username = "user";
-        String password = "123";
-        when(personService.getPerson(username)).thenReturn(new Person(1, username, password, PersonRoleEnum.ADMIN));
-        when(person.getRole()).thenReturn(PersonRoleEnum.ADMIN);
-        roles.add(new SimpleGrantedAuthority(person.getRole().getEnumrole()));
-        UserDetails expectedPerson = new User(username, password, roles);
-        UserDetails actualPerson = personDetailsServiceImpl.loadUserByUsername(username);
-        assertEquals(expectedPerson.getUsername(), actualPerson.getUsername());
+    void getPersonPositiveResult() {
+        String name = "user";
+        when(personDAO.getPersons()).thenReturn(personList);
+        Person actualPerson = personService.getPerson(name);
+        assertEquals(expectedPerson, actualPerson);
     }
 
     @Test
-    void loadUserThrowsException() {
-        String nonExistUsername = "1111";
-        when(personService.getPerson(nonExistUsername)).thenThrow(UsernameNotFoundException.class);
-        when(person.getRole()).thenReturn(PersonRoleEnum.ADMIN);
-        roles.add(new SimpleGrantedAuthority(person.getRole().getEnumrole()));
-        try {
-            personDetailsServiceImpl.loadUserByUsername(nonExistUsername);
-            fail();
-        } catch (UsernameNotFoundException thrown) {
-            assertNotNull(thrown);
-        }
+    void getPersonNotFound() {
+        String name = "tmp";
+        when(personDAO.getPersons()).thenReturn(personList);
+        Person actualPerson = personService.getPerson(name);
+        assertNull(actualPerson);
     }
 }
